use int for generate_random_bbox

+++ python/coral_usb/util.py
@@ -94,5 +94,5 @@ def generate_random_bbox(n, img_size, min_length, max_length):
     x_min = np.random.uniform(0, W - max_length, size=(n,))
     y_max = y_min + np.random.uniform(min_length, max_length, size=(n,))
     x_max = x_min + np.random.uniform(min_length, max_length, size=(n,))
-    bbox = np.stack((y_min, x_min, y_max, x_max), axis=1).astype(np.float32)
+    bbox = np.stack((y_min, x_min, y_max, x_max), axis=1).astype(np.int)
     return bbox
